backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
redis>=5.0.0
rq>=1.16.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
Pillow>=10.2.0
openai>=1.12.0
google-generativeai>=0.4.0
websockets>=12.0
python-multipart>=0.0.6
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg[binary]>=3.1.18
alembic>=1.13.0
jinja2>=3.1.3
